requirements.txt
Flask==3.0.3
gunicorn==21.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
flask-cors==5.0.0
